refactor(card_inspector_panel): split CardInspectorPanel into mixin package
Move widgets/panels/card_inspector_panel.py into a package: frame.py
builds the UI; properties.py holds the image-request name resolver,
match checker, and failure-key helper; handlers.py owns reset,
update_card, printings navigation, async loaders, and the display-mode
logic.
