+++ widgets/panels/card_inspector_panel/__init__.py
@@ -1,0 +1,7 @@
+"""Card inspector panel package."""
+
+from __future__ import annotations
+
+from widgets.panels.card_inspector_panel.frame import CardInspectorPanel
+
+__all__ = ["CardInspectorPanel"]
